+++ conftest.py
@@ -1,0 +1,3 @@
+# A hack to force the inclusion of the 'subscriber' moduls for pytest
+import sys, os
+sys.path.append(os.getcwd()+"/subscriber")