+++ subscriber/podaac_data_downloader.py
@@ -94,7 +94,7 @@ def run():
 
     if not isdir(data_path):
         print("NOTE: Making new data directory at " + data_path + "(This is the first run.)")
-        makedirs(data_path)
+        makedirs(data_path, exist_ok=True)
 
     # Change this to whatever extent you need. Format is W Longitude,S Latitude,E Longitude,N Latitude
     bounding_extent = args.bbox