+++ CHANGELOG.md
@@ -12,7 +12,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - created library of common access mechanisms to split between subscriber and downloader capabilities
 - added .tar.gz to list of default extensions. [40](https://github.com/podaac/data-subscriber/issues/40)
 - Ignore error if destination directory already exists. [46](https://github.com/podaac/data-subscriber/issues/46)
+- Updated the naming convention of .update file. [44](https://github.com/podaac/data-subscriber/issues/44)
 ### Deprecated
+- use of ".update" file naming convention. This will still work, but will be renamed to .update__COLLECTIONNAME after a successful run. the ".update" file will need to be manually cleaned up. See [issue 44](https://github.com/podaac/data-subscriber/issues/44)
 ### Removed
 ### Fixed
 - issue where only specifying an end data cause issues in subscriber. [39](https://github.com/podaac/data-subscriber/issues/39)